Refactor PDF processing into PdfProcessor class
Moved PDF-to-image conversion logic from main.ts to PdfProcessor.ts class for better separation of concerns.
Updated main.ts to use PdfProcessor for handling PDF files.
